telegram-bot/requirements.txt
python-telegram-bot==12.8
paho-mqtt==1.5.0